+++ changelog.md
@@ -10,6 +10,7 @@
 - Added visual indicators for when game start options or meta tree weapons are locked behind an unobtained Archipelago item
 - Added visual indicators for unchecked locations (chests/perks, weapons, characters)
 - Fixed incorrect characters being identified for unlock/escape conditions
+- Added client config options to automatically connect to Archipelago on game launch
 - Backend:
   - The client will no longer even attempt to send checks that aren't active/valid on the server
   - Fixed a probably harmless error that could sometimes occur because skill tree itemization was trying to activate on the meta tree